my_app/requirements.txt
aioredis==1.3.1
alabaster==0.7.12
amqp==5.0.9
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.1
django-allauth==0.47.0
django-autoslug==1.9.8
django-ckeditor==6.2.0
django-constance==2.8.0
django-debug-toolbar==3.2.4
django-js-asset==1.2.2
django-phonenumber-field==6.0.0
django-picklefield==3.0.1
django-redis==5.2.0
django-stubs==1.9.0
django-stubs-ext==0.3.1
docutils==0.17.1
factory-boy==3.2.1
Faker==11.3.0
flake8==3.9.2
flower==1.0.0
h11==0.12.0
hiredis==2.0.0
humanize==3.13.1
hyperlink==21.0.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.10.1
incremental==21.3.0
iniconfig==1.1.1
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.3
mypy==0.931
mypy-extensions==0.4.3
oauthlib==3.1.1
outcome==1.1.0
packaging==21.3
phonenumbers==8.12.40
Pillow==8.4.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
Pygments==2.11.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
python-dateutil==2.8.2

python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
redis==4.0.2
requests==2.27.0
requests-oauthlib==1.3.0
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sorl-thumbnail==12.7.0
sortedcontainers==2.4.0
speedtest-cli==2.1.3
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.0
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
twilio==7.4.0
Twisted==21.7.0
txaio==21.2.1
types-pytz==2021.3.4
types-PyYAML==6.0.3
typing-extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
wsproto==1.0.0
zipp==3.7.0
zope.interface==5.4.0

djangorestframework~=3.13.1